Move input/ouput validation to a shared util
