--- benchmarks/src/duckdb_utils.rs
@@ -1,42 +1,0 @@
-// Copyright (c) 2023-2026 ParadeDB, Inc.
-//
-// This file is part of ParadeDB - Postgres for Search and Analytics
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-use anyhow::{Context, Result};
-use duckdb::{AccessMode, Config, Connection};
-
-pub fn open_duckdb_conn() -> Result<Connection> {
-    let config = Config::default()
-        .access_mode(AccessMode::Automatic)?
-        .enable_autoload_extension(true)?;
-    let conn = Connection::open_in_memory_with_flags(config)
-        .context("Failed to open DuckDB in-memory connection")?;
-
-    conn.execute_batch(
-        "CREATE OR REPLACE SECRET secret (TYPE s3, PROVIDER credential_chain);",
-    )
-    .context("Failed to configure S3 credentials. Ensure AWS credentials are available via environment variables, ~/.aws/credentials, or instance metadata.")?;
-
-    conn.execute("INSTALL httpfs", [])
-        .with_context(|| "Failed to install httpfs extension")?;
-    conn.execute("LOAD httpfs", [])
-        .with_context(|| "Failed to load httpfs extension")?;
-    // Increase timeout (default is 30 seconds) to allow for working with larger files (200MB+)
-    conn.execute("SET http_timeout = 120", [])
-        .with_context(|| "Failed to configure http timeout")?;
-
-    Ok(conn)
-}
